+++ ligo/skymap/util/tests/test_progress.py
@@ -4,12 +4,17 @@ from time import sleep
 import numpy as np
 import pytest
 
-from ..progress import progress_map
+from ..progress import progress_map, progress_map_vectorized
 
 
 def func(x):
     sleep(np.random.uniform(0, 0.1))
     return np.square(x)
+
+
+def func2(x):
+    sleep(np.random.uniform(0, 0.1))
+    return np.cos(x), np.sin(x)
 
 
 @pytest.mark.parametrize("jobs", [1, 8, None])
@@ -40,3 +45,18 @@ def test_no_nested_pools():
 def test_indefinite(jobs):
     """Test iteration over a collection of indefinite length."""
     assert list(progress_map(np.square, (i for i in range(3)), jobs=jobs)) == [0, 1, 4]
+
+
+@pytest.mark.parametrize("jobs", [1, 8, None])
+def test_map_vectorized(jobs):
+    x = np.arange(20)
+    result = progress_map_vectorized(func, x, jobs=jobs)
+    np.testing.assert_array_equal(result, np.square(x))
+
+
+@pytest.mark.parametrize("jobs", [1, 8, None])
+def test_map_vectorized_nout(jobs):
+    x = np.arange(20)
+    result1, result2 = progress_map_vectorized(func2, x, jobs=jobs, nout=2)
+    np.testing.assert_array_equal(result1, np.cos(x))
+    np.testing.assert_array_equal(result2, np.sin(x))